Remove AC_FUNC_MALLOC
If using this, one is supposed to provide a malloc replacement if
malloc isn't found. In practice, AC_FUNC_MALLOC always fails if
cross compiling.

When a config.h is generated, as in the opencore-amr 0.1.3 release,
this didn't matter (since nothing actually included config.h), but
now -Dmalloc=rpl_malloc gets passed to the compilation steps instead.

Therefore remove this test - we're actually not interested in the
semantics of malloc(0) at all.